adding user endpoints to api

+++ server/controllers/userController.js
@@ -6,7 +6,14 @@ exports.getUsers = async (req, res) => {
   res.json(users)
 };
 
-exports.getAuthUser = () => {};
+exports.getAuthUser = (req, res) => {
+  if(!req.isAuthUser){
+    return res.status(403).json({
+      message: "You are unauthenticated. Please sign in or sign up."
+    })
+  }
+  res.json(req.user)
+};
 
 exports.getUserById = async (req, res, next, id) => {
   const user = await User.findOne({ _id: id })
@@ -21,7 +28,14 @@ exports.getUserById = async (req, res, next, id) => {
   next()
 };
 
-exports.getUserProfile = () => {};
+exports.getUserProfile = (req, res) => {
+  if(!req.profile){
+    return res.status(404).json({
+      message: "No user found."
+    })
+  }
+  return res.json(req.profile)
+};
 
 exports.getUserFeed = () => {};
 
@@ -33,8 +47,12 @@ exports.updateUser = () => {};
 
 exports.deleteUser = async (req, res) => {
   const { userId } = req.params;
+  if(!req.isAuthUser){
+    return res.status(400).json({ message: 'You are not authorized to perform this action.' })
+  }
 
   const deletedUser = await User.findOneAndDelete({ _id: userId })
+  res.json(deletedUser)
 };
 
 exports.addFollowing = () => {};
